refactor: 플랫폼 동기화 시작 로직을 훅으로 이동
